Split out old version and converters
