+++ qcodes/dataset/descriptions/versioning/v0.py
@@ -1,0 +1,52 @@
+from typing import Dict, Any
+
+from qcodes.dataset.descriptions.param_spec import ParamSpec
+
+
+class InterDependencies:
+    """
+    Object containing the ParamSpecs of a given run
+    """
+
+    def __init__(self, *paramspecs: ParamSpec) -> None:
+
+        for paramspec in paramspecs:
+            if not isinstance(paramspec, ParamSpec):
+                raise ValueError('Got invalid input. All paramspecs must be '
+                                 f'ParamSpecs, but {paramspec} is of type '
+                                 f'{type(paramspec)}.')
+
+        self.paramspecs = paramspecs
+
+    def __repr__(self) -> str:
+        output = self.__class__.__name__
+        tojoin = (str(paramspec) for paramspec in self.paramspecs)
+        output += f'({", ".join(tojoin)})'
+        return output
+
+    def __eq__(self, other) -> bool:
+        if not isinstance(other, InterDependencies):
+            return False
+        ours = sorted(self.paramspecs, key=lambda ps: ps.name)
+        theirs = sorted(other.paramspecs, key=lambda ps: ps.name)
+        if not ours == theirs:
+            return False
+        return True
+
+    def serialize(self) -> Dict[str, Any]:
+        """
+        Return a serialized version of this object instance
+        """
+        ser = {}
+        ser['paramspecs'] = tuple(ps.serialize() for ps in self.paramspecs)
+        return ser
+
+    @classmethod
+    def deserialize(cls, ser: Dict[str, Any]) -> 'InterDependencies':
+        """
+        Create an InterDependencies object from a serialization of an
+        instance
+        """
+        paramspecs = [ParamSpec.deserialize(sps) for sps in ser['paramspecs']]
+        idp = cls(*paramspecs)
+        return idp
