+++ qcodes/dataset/sqlite/db_upgrades/upgrade_2_to_3.py
@@ -8,8 +8,8 @@ from tqdm import tqdm
 from qcodes.dataset.sqlite.connection import ConnectionPlus, transaction, \
     atomic, atomic_transaction
 from qcodes.dataset.sqlite.query_helpers import one
-from qcodes.dataset.param_spec import ParamSpec
-from qcodes.dataset.dependencies import InterDependencies
+from qcodes.dataset.descriptions.param_spec import ParamSpec
+from qcodes.dataset.descriptions.versioning.v0 import InterDependencies
 
 
 log = logging.getLogger(__name__)
