+++ qcodes/dataset/sqlite/db_upgrades/upgrade_5_to_6.py
@@ -2,7 +2,7 @@ import json
 
 from tqdm import tqdm
 
-from qcodes.dataset.dependencies import InterDependencies
+from qcodes.dataset.descriptions.versioning.v0 import InterDependencies
 from qcodes.dataset.sqlite.connection import ConnectionPlus, atomic, \
     atomic_transaction
 from qcodes.dataset.sqlite.queries import get_run_description, \
