+++ qcodes/dataset/sqlite/queries.py
@@ -15,15 +15,16 @@ from typing import Dict, List, Optional, Any, Sequence, Union, Tuple, \
 import numpy as np
 
 import qcodes as qc
+from qcodes.dataset.descriptions.rundescriber import RunDescriber
+from qcodes.dataset.descriptions.param_spec import ParamSpec
+from qcodes.dataset.descriptions.versioning.converters import old_to_new
+from qcodes.dataset.descriptions.versioning.v0 import InterDependencies
 from qcodes.dataset.guids import parse_guid, generate_guid
 from qcodes.dataset.sqlite.connection import transaction, ConnectionPlus, \
     atomic_transaction, atomic
 from qcodes.dataset.sqlite.query_helpers import sql_placeholder_string, \
     many_many, one, many, select_one_where, select_many_where, insert_values, \
     insert_column, VALUES, update_where
-from qcodes.dataset.dependencies import InterDependencies, old_to_new
-from qcodes.dataset.descriptions import RunDescriber
-from qcodes.dataset.param_spec import ParamSpec
 
 
 log = logging.getLogger(__name__)
