+++ qcodes/tests/dataset_generators.py
@@ -1,6 +1,6 @@
 import numpy as np
-from qcodes.dataset.param_spec import ParamSpecBase
-from qcodes.dataset.dependencies import InterDependencies_
+from qcodes.dataset.descriptions.param_spec import ParamSpecBase
+from qcodes.dataset.descriptions.dependencies import InterDependencies_
 
 
 def dataset_with_outliers_generator(ds, data_offset=5, low_outlier=-3,
